Move constants and functions in dx-grid to adjacent `util.ts`.
